feat(article): finish implment article crud except update (edit)

+++ src/server/api/routers/article.router.ts
@@ -1,10 +1,15 @@
 import {
-  createArticleSchema, deleteArticleSchema,
-  getArticleSchema, togglePublishArticleSchema,
+  createArticleSchema,
+  deleteArticleSchema,
+  deleteArticlesSchema,
+  getArticleSchema,
+  getArticlesSchema,
+  togglePublishArticleSchema,
+  togglePublishArticlesSchema,
 } from "../schemas/article.schema";
 import { adminProcedure, createTRPCRouter, publicProcedure } from "../trpc";
-import {TRPCError} from "@trpc/server";
-import {UTApi} from "uploadthing/server";
+import { TRPCError } from "@trpc/server";
+import { UTApi } from "uploadthing/server";
 
 const utapi = new UTApi();
 
@@ -24,13 +29,50 @@ export const articleRouter = createTRPCRouter({
       },
     });
   }),
-  getArticles: publicProcedure.query(async ({ ctx }) => {
-    return ctx.db.article.findMany({
-      where: {
-        isPublished: true,
-      },
-    });
-  }),
+  getArticles: publicProcedure
+    .input(getArticlesSchema)
+    .query(async ({ ctx, input }) => {
+      const { search, orderBy, orderDirection } = input;
+      // default = { search: '', orderBy: 'publishedAt', orderDirection: 'desc' }
+      const isDefaultSearch =
+        search === "" && orderBy === "publishedAt" && orderDirection === "desc";
+
+      const articles = await ctx.db.article.findMany({
+        where: {
+          isPublished: true,
+          title: {
+            contains: search,
+          },
+        },
+        orderBy: {
+          publishedAt: orderBy === "publishedAt" ? orderDirection : undefined,
+          updatedAt: orderBy === "updatedAt" ? orderDirection : undefined,
+          title: orderBy === "title" ? orderDirection : undefined,
+          views: orderBy === "views" ? orderDirection : undefined,
+          likes: orderBy === "likes" ? orderDirection : undefined,
+        },
+        include: {
+          author: {
+            select: {
+              image: true,
+              name: true,
+            },
+          },
+        },
+      });
+
+      const isNotFound = isDefaultSearch && articles.length === 0;
+
+      if (isNotFound) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message:
+            "Artikel tidak ditemukan, tunggu hingga artikel dipublikasikan",
+        });
+      }
+
+      return articles;
+    }),
   getArticle: publicProcedure
     .input(getArticleSchema)
     .query(async ({ ctx, input }) => {
@@ -45,11 +87,20 @@ export const articleRouter = createTRPCRouter({
             },
           ],
         },
+        include: {
+          author: {
+            select: {
+              image: true,
+              name: true,
+            },
+          },
+        },
       });
     }),
   createArticle: adminProcedure
     .input(createArticleSchema)
     .mutation(async ({ ctx, input }) => {
+      console.log(input);
       return ctx.db.article.create({
         data: {
           ...input,
@@ -70,17 +121,51 @@ export const articleRouter = createTRPCRouter({
       if (!article) {
         throw new TRPCError({
           code: "NOT_FOUND",
-            message: "Artikel tidak ditemukan",
-        })
+          message: "Artikel tidak ditemukan",
+        });
       }
 
+      const publishedAt = input.isPublished ? new Date() : null;
       return ctx.db.article.update({
         where: {
           id: input.id,
         },
         data: {
           isPublished: input.isPublished,
-          publishedAt: input.isPublished ? new Date() : null,
+          publishedAt,
+        },
+      });
+    }),
+  togglePublishArticles: adminProcedure
+    .input(togglePublishArticlesSchema)
+    .mutation(async ({ ctx, input }) => {
+      const { ids, isPublished } = input;
+
+      const articles = await ctx.db.article.findMany({
+        where: {
+          id: {
+            in: ids,
+          },
+        },
+      });
+
+      if (articles.length !== ids.length) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Artikel tidak ditemukan",
+        });
+      }
+
+      const publishedAt = isPublished ? new Date() : null;
+      return ctx.db.article.updateMany({
+        where: {
+          id: {
+            in: ids,
+          },
+        },
+        data: {
+          isPublished,
+          publishedAt,
         },
       });
     }),
@@ -88,24 +173,75 @@ export const articleRouter = createTRPCRouter({
     .input(deleteArticleSchema)
     .mutation(async ({ ctx, input }) => {
       const article = await ctx.db.article.findUnique({
-          where: {
-              id: input.id,
-          },
+        where: {
+          id: input.id,
+        },
       });
 
       if (!article) {
-          throw new TRPCError({
-              code: "NOT_FOUND",
-              message: "Artikel tidak ditemukan",
-          })
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Artikel tidak ditemukan",
+        });
       }
 
-      await utapi.deleteFiles([article.thumbnailKey]);
+      if (article.thumbnailKey) {
+        await utapi.deleteFiles([article.thumbnailKey]);
+      }
 
       return ctx.db.article.delete({
-          where: {
-              id: input.id,
+        where: {
+          id: input.id,
+        },
+      });
+    }),
+  deleteArticles: adminProcedure
+    .input(deleteArticlesSchema)
+    .mutation(async ({ ctx, input }) => {
+      const articles = await ctx.db.article.findMany({
+        where: {
+          id: {
+            in: input.ids,
           },
+        },
+      });
+
+      if (articles.length !== input.ids.length) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Artikel tidak ditemukan",
+        });
+      }
+
+      const thumbnailKeys = articles
+        .filter((article) => article.thumbnailKey)
+        .map((article) => article.thumbnailKey)
+        .filter((key): key is string => key !== null);
+
+      if (thumbnailKeys.length) {
+        await utapi.deleteFiles(thumbnailKeys);
+      }
+
+      return ctx.db.article.deleteMany({
+        where: {
+          id: {
+            in: input.ids,
+          },
+        },
+      });
+    }),
+  incrementView: publicProcedure
+    .input(getArticleSchema)
+    .mutation(async ({ ctx, input }) => {
+      return ctx.db.article.update({
+        where: {
+          id: input.id,
+        },
+        data: {
+          views: {
+            increment: 1,
+          },
+        },
       });
     }),
 });
